AI_Employee_Vault/Skills/File_System_Watcher/requirements.txt
# File System Watcher — Bronze Tier
python-dotenv>=1.0.0
watchdog>=4.0.0
